requirements.txt
streamlit==1.32.0
google-generativeai==0.4.1
python-dotenv==1.0.1
PyPDF2==3.0.1
Pillow==10.2.0
anyio==4.3.0 
packaging==23.2 
tenacity==8.2.3 
docutils==0.20.1